pyproject.toml
[tool.poetry]
name = "hand_gesture_regonition"
version = "0.1.0"
description = ""
authors = ["Will Meyburgh <annekinmeyburgh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.11.0.86"
mediapipe = "^0.10.21"
torch = {version = "^2.7.0+cpu", source = "pytorch-cpu"}
torchvision = {version = "^0.22.0+cpu", source = "pytorch-cpu"}
python-dotenv = "^1.1.0"
homeassistant-api = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"